+++ src/typedal/mixins.py
@@ -11,6 +11,8 @@ import warnings
 from datetime import datetime
 from typing import Any, Optional
 
+from pydal import DAL
+from pydal.validators import IS_NOT_IN_DB, ValidationError
 from slugify import slugify
 
 from .core import (  # noqa F401 - used by example in docstring
@@ -87,6 +89,56 @@ def slug_random_suffix(length: int = 8) -> str:
     return base64.urlsafe_b64encode(os.urandom(length)).rstrip(b"=").decode().strip("=")
 
 
+T = typing.TypeVar("T")
+
+
+# noinspection PyPep8Naming
+class HAS_UNIQUE_SLUG(IS_NOT_IN_DB):
+    """
+    Checks if slugified field is already in the db.
+
+    Usage:
+        table.name = HAS_UNIQUE_SLUG(db, "table.slug")
+    """
+
+    def __init__(
+        self,
+        db: TypeDAL | DAL,
+        field: str,  # table.slug
+        error_message: str = "This slug is not unique: %s.",
+    ):
+        """
+        Based on IS_NOT_IN_DB but with less options and a different default error message.
+        """
+        super().__init__(db, field, error_message)
+
+    def validate(self, original: T, record_id: Optional[int] = None) -> T:
+        """
+        Performs checks to see if the slug already exists for a different row.
+        """
+        value = slugify(str(original))
+        if not value.strip():
+            raise ValidationError(self.translator(self.error_message))
+
+        (tablename, fieldname) = str(self.field).split(".")
+        table = self.dbset.db[tablename]
+        field = table[fieldname]
+        query = field == value
+
+        # make sure exclude the record_id
+        row_id = record_id or self.record_id
+        if isinstance(row_id, dict):  # pragma: no cover
+            row_id = table(**row_id)
+        if row_id is not None:
+            query &= table._id != row_id
+        subset = self.dbset(query)
+
+        if subset.count():
+            raise ValidationError(self.error_message % value)
+
+        return original
+
+
 class SlugMixin(Mixin):
     """
     (Opinionated) example mixin to add a 'slug' field, which depends on a user-provided other field.
@@ -140,25 +192,39 @@ class SlugMixin(Mixin):
         cls.__settings__["slug_suffix"] = slug_suffix
 
     @classmethod
+    def __generate_slug_before_insert(cls, row: OpRow) -> None:
+        settings = cls.__settings__
+
+        text_input = row[settings["slug_field"]]
+        generated_slug = slugify(text_input)
+
+        if suffix_len := settings["slug_suffix"]:
+            generated_slug += f"-{slug_random_suffix(suffix_len)}"
+
+        row["slug"] = slugify(generated_slug)
+        return None
+
+    @classmethod
     def __on_define__(cls, db: TypeDAL) -> None:
         """
         When db is available, include a before_insert hook to generate and include a slug.
         """
         super().__on_define__(db)
+        settings = cls.__settings__
 
         # slugs should not be editable (for SEO reasons), so there is only a before insert hook:
-        def generate_slug_before_insert(row: OpRow) -> None:
-            settings = cls.__settings__
+        cls._before_insert.append(cls.__generate_slug_before_insert)
 
-            text_input = row[settings["slug_field"]]
-            generated_slug = slugify(text_input)
+        if settings["slug_suffix"] == 0:
+            # add a validator to the field that will be slugified:
+            slug_field = getattr(cls, settings["slug_field"])
+            current_requires = getattr(slug_field, "requires", None) or []
+            if not isinstance(current_requires, list):
+                current_requires = [current_requires]
 
-            if suffix_len := settings["slug_suffix"]:
-                generated_slug += f"-{slug_random_suffix(suffix_len)}"
+            current_requires.append(HAS_UNIQUE_SLUG(db, f"{cls}.slug"))
 
-            row["slug"] = slugify(generated_slug)
-
-        cls._before_insert.append(generate_slug_before_insert)
+            slug_field.requires = current_requires
 
     @classmethod
     def from_slug(cls: typing.Type[T_MetaInstance], slug: str, join: bool = True) -> Optional[T_MetaInstance]: